% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependecies.R
\name{add_details}
\alias{add_details}
\title{Check a dependency list for inclusion in the lockfile and add those additional
details (repo, remote, version, subdir) if it does appear there.  Also throw
an error if we require a dependency version greater than that specified by
the lockfile.}
\usage{
add_details(current_list, lock)
}
\description{
Check a dependency list for inclusion in the lockfile and add those additional
details (repo, remote, version, subdir) if it does appear there.  Also throw
an error if we require a dependency version greater than that specified by
the lockfile.
}

