% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependecies.R
\name{add_latest_version_in_lockbox}
\alias{add_latest_version_in_lockbox}
\title{Attach the latest available lockbox version to a package}
\usage{
add_latest_version_in_lockbox(package)
}
\description{
Attach the latest available lockbox version to a package
}

