% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{age_decrypt_key}
\alias{age_decrypt_key}
\title{Decrypt an age-encrypted file using a private key}
\usage{
age_decrypt_key(encrypted_file_path, private_key_path)
}
\description{
Decrypt an age-encrypted file using a private key
}
