% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{age_decrypt_passphrase}
\alias{age_decrypt_passphrase}
\title{Decrypt an age-encrypted file using a passphrase}
\usage{
age_decrypt_passphrase(encrypted_file_path, passphrase)
}
\description{
Decrypt an age-encrypted file using a passphrase
}
