% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{age_encrypt}
\alias{age_encrypt}
\title{Encrypt a file using age}
\usage{
age_encrypt(
  input = NULL,
  output = if (!is.null(input)) paste0(input, ".age") else NULL,
  public = NULL,
  overwrite = FALSE,
  armor = FALSE
)
}
\description{
Encrypt a file using age
}
