% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dependecies.R
\name{combine_dependencies}
\alias{combine_dependencies}
\title{Combine two lists of dependencies via version comparisons.  Keep packages
found in list1 on the left side of the entirety of list2, while moving
the parent package to the space after it's rightmost dependency found in list2.}
\usage{
combine_dependencies(list1, list2, current_parent)
}
\description{
Combine two lists of dependencies via version comparisons.  Keep packages
found in list1 on the left side of the entirety of list2, while moving
the parent package to the space after it's rightmost dependency found in list2.
}

