% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staging.R
\name{copy_real_packages_to_lockbox_library}
\alias{copy_real_packages_to_lockbox_library}
\title{Copy real (non-symlinked) packages to the lockbox library.}
\usage{
copy_real_packages_to_lockbox_library(staging_library)
}
\arguments{
\item{staging_library}{character. The location of the staging library.}
}
\description{
Copy real (non-symlinked) packages to the lockbox library.
}
