% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{current_version}
\alias{current_version}
\title{The current version of this package in the current library.}
\usage{
current_version(pkg)
}
\arguments{
\item{pkg}{character or locked_package. The name of the package.}
}
\value{
a \code{\link{package_version}} object representing the version of
  this package in the current library.
}
\description{
The current version of this package in the current library.
}
