% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dependecies.R
\name{dependencies_from_description}
\alias{dependencies_from_description}
\title{Parse dependencies from description}
\usage{
dependencies_from_description(package, dcf)
}
\description{
Parse dependencies from description
}

