% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dependecies.R
\name{download_package.CRAN}
\alias{download_package.CRAN}
\title{Download CRAN package, either current or older version}
\usage{
\method{download_package}{CRAN}(package)
}
\description{
Download CRAN package, either current or older version
}

