% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependecies.R
\name{download_package.github}
\alias{download_package.github}
\title{Download a package from github our version of devtools' remote_download
function}
\usage{
\method{download_package}{github}(package)
}
\description{
Download a package from github our version of devtools' remote_download
function
}

