% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{ensure_package_exists_in_lockbox}
\alias{ensure_package_exists_in_lockbox}
\alias{ensure_package_exists_in_lockbox!}
\title{Ensure package name and version exists in the lockbox secret library.}
\usage{
"ensure_package_exists_in_lockbox!"(locked_package)
}
\arguments{
\item{locked_package}{locked_package. In particular, \code{name} and
\code{version} elements will be used. If the package version is
not present in the lockbox, we will attempt to download it from
CRAN or github.}
}
\description{
The lockbox package keeps a secret directory with the packages for
each given version. By default, this is in
\code{getOption("lockbox.lib")} or \code{"~/.R/lockbox"} if that option
is not set.
}
\note{
The weird name is a derivative of a \href{http://stackoverflow.com/questions/612189/why-are-exclamation-marks-used-in-ruby-methods}{Rubyism}
   to indicate some serious side effects can occur! In this case, we
   download the package from CRAN or github if the name + version combo
   does not exist.
}
