% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{get_available_cran_version}
\alias{get_available_cran_version}
\title{Return the latest available version of a package from CRAN}
\usage{
get_available_cran_version(package, repo = "http://cran.r-project.org")
}
\description{
Return the latest available version of a package from CRAN
}

