% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependecies.R
\name{get_dependencies}
\alias{get_dependencies}
\title{Either use the current lockbox library DESCRIPTION
file or download the accurate remote DESCRIPTION file.}
\usage{
get_dependencies(package, lock)
}
\description{
Either use the current lockbox library DESCRIPTION
file or download the accurate remote DESCRIPTION file.
}

