% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependecies.R
\name{get_dependencies_for_list}
\alias{get_dependencies_for_list}
\title{Recursive function to take a list and lock and extract dependencies, sorting
along the way using the combine_dependencies function.}
\usage{
get_dependencies_for_list(master_list, lock, previously_parsed_deps,
  current_parent)
}
\description{
Recursive function to take a list and lock and extract dependencies, sorting
along the way using the combine_dependencies function.
}

