% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependecies.R
\name{get_ordered_dependencies}
\alias{get_ordered_dependencies}
\title{Get dependencies for all elements with lock, but only do so for current
version mismatches and non-cran installations}
\usage{
get_ordered_dependencies(lock, mismatches)
}
\description{
Get dependencies for all elements with lock, but only do so for current
version mismatches and non-cran installations
}

