% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependecies.R
\name{get_remote_dependencies.CRAN}
\alias{get_remote_dependencies.CRAN}
\title{For packages on CRAN we will extract to a temporary directory when we
download the accurate remote DESCRIPTION file.}
\usage{
\method{get_remote_dependencies}{CRAN}(package)
}
\description{
For packages on CRAN we will extract to a temporary directory when we
download the accurate remote DESCRIPTION file.
}

