% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dependecies.R
\name{get_remote_dependencies.github}
\alias{get_remote_dependencies.github}
\title{Download the accurate remote DESCRIPTION file for a github repo.}
\usage{
\method{get_remote_dependencies}{github}(package)
}
\description{
Download the accurate remote DESCRIPTION file for a github repo.
}

