% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dependecies.R
\name{get_remote_dependencies.local}
\alias{get_remote_dependencies.local}
\title{If a package is local we just read from the directory given}
\usage{
\method{get_remote_dependencies}{local}(package)
}
\description{
If a package is local we just read from the directory given
}

