% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dependecies.R
\name{get_remote_dependencies}
\alias{get_remote_dependencies}
\title{Get the dependencies for a given package}
\usage{
get_remote_dependencies(package)
}
\description{
Get the dependencies for a given package
}

