% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependecies.R
\name{github_remote}
\alias{github_remote}
\title{Get remote for a repo given a username, ref, subdir, auth_token, etc.  Stolen
shamelessly from devtools}
\usage{
github_remote(repo, username = NULL, ref = NULL, subdir = NULL,
  auth_token = github_pat(), sha = NULL, host = "api.github.com")
}
\description{
Get remote for a repo given a username, ref, subdir, auth_token, etc.  Stolen
shamelessly from devtools
}

