% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{install_package.github}
\alias{install_package.github}
\title{Install package by downloading from github}
\usage{
\method{install_package}{github}(locked_package)
}
\description{
Install package by downloading from github
}

