% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lockbox.R, R/package.lockbox.R
\docType{package}
\name{lockbox}
\alias{lockbox}
\alias{lockbox-package}
\title{Re-organize Search Path to Use Lockbox Library.}
\usage{
lockbox(file_or_list)
}
\arguments{
\item{file_or_list}{character or list. A yaml-based lock file or its
parsed out list format. This set of packages will be loaded into the
search path and \emph{all other packages will be unloaded}.}
}
\description{
The lockbox package provides a separate directory, by default under
\code{"~/.R/.lockbox"} (although this is configurable from the global ption
\code{"lockbox.directory"}) that maintains different versions of packages
on demand. When a given set of versioned packages is requested, lockbox will
unload \emph{all other packages} and ensure only the given set of packages
with their respective versions are present.

Bundler-style Package Management for R.
}
\details{
Since lockbox maintains a separate directory for its library, it will not
interfere with R's usual packages or libraries when R is restarted.
}

