% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lockbox.R
\name{lockbox_transient_staging_dir}
\alias{lockbox_transient_staging_dir}
\title{The transient staging lockbox library path.}
\usage{
lockbox_transient_staging_dir()
}
\description{
This will be used to copy interactively installed packages to
the vanilla library.
}

