% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dependecies.R
\name{max_package_version}
\alias{max_package_version}
\title{Take a vector of versions and find the max without coercing to package_version}
\usage{
max_package_version(versions)
}
\description{
Take a vector of versions and find the max without coercing to package_version
}

