% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{namespace_loading_install_github}
\alias{namespace_loading_install_github}
\title{Function to load devtools package, then unload for installing from github.
Necessary for remotes to be correctly parsed by devtools :/}
\usage{
namespace_loading_install_github(...)
}
\description{
Function to load devtools package, then unload for installing from github.
Necessary for remotes to be correctly parsed by devtools :/
}

