% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{package_list}
\alias{package_list}
\title{Recursive function to take a list and lock and extract dependencies, sorting
along the way using the combine_dependencies function.}
\usage{
package_list(master_list, lock, previously_parsed_deps)
}
\arguments{
\item{master_list}{list.  Packages to go through and parse dependencies from}

\item{lock}{list.  Original list of locked packages}

\item{previously_parsed_deps}{list.  List of packages and their dependencies that
we have already parsed out of their respective description files}
}
\description{
Recursive function to take a list and lock and extract dependencies, sorting
along the way using the combine_dependencies function.
}

