% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dependecies.R
\name{parse_dcf}
\alias{parse_dcf}
\title{Borrowed from tools::package.dependencies and modified to be less breaky
and parse remotes}
\usage{
parse_dcf(x, check = FALSE, depLevel = c("Depends", "Imports", "Suggests",
  "Remotes"))
}
\description{
Borrowed from tools::package.dependencies and modified to be less breaky
and parse remotes
}

