% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebuild.R
\name{rebuild}
\alias{rebuild}
\title{Rebuild our current search path according to a list of locked packages.}
\usage{
rebuild(packages)
}
\arguments{
\item{packages}{list of locked_packages.}
}
\description{
As with the \href{https://github.com/romainfrancois/nothing}{nothing package},
this function will unload everything except the base R packages from the
search path prior to re-building.
}

