% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependecies.R
\name{remote_download_github_remote}
\alias{remote_download_github_remote}
\title{Download a package using the github remote arguments.  Borrowed entirely from
devtools}
\usage{
remote_download_github_remote(x, quiet = FALSE)
}
\description{
Download a package using the github remote arguments.  Borrowed entirely from
devtools
}

