% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependecies.R
\name{replace_with_lock}
\alias{replace_with_lock}
\title{Check a dependency list for inclusion in the lockfile and replace the package
with the locked version if it does appear there.  Also throw
an error if we require a dependency version greater than that specified by
the lockfile.}
\usage{
replace_with_lock(package, lock)
}
\description{
Check a dependency list for inclusion in the lockfile and replace the package
with the locked version if it does appear there.  Also throw
an error if we require a dependency version greater than that specified by
the lockfile.
}

