% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sops.R
\name{sops_recipients}
\alias{sops_recipients}
\title{Extract age recipients from encrypted SOPS file}
\usage{
sops_recipients(lockbox)
}
\arguments{
\item{lockbox}{Character string, path to the SOPS-encrypted file}
}
\value{
Character vector of age recipient public keys
}
\description{
Reads a SOPS-encrypted file and extracts the age recipient public keys
that can decrypt the file.
}
\keyword{internal}
