% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staging.R
\name{staging_library}
\alias{staging_library}
\title{Construct a lockbox staging library.}
\usage{
staging_library()
}
\value{
the directory of the staging library. This should be unlinked 
  after use.
}
\description{
To install a new package to the lockbox library, a virtual library is
constructed that is used during the installation process. This
function generates that library and returns its (temporary)
directory name. See the vignette on "Libraries in Lockbox" for more details.
}

