% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dependecies.R
\name{strip_duplicate_dependencies}
\alias{strip_duplicate_dependencies}
\title{We don't trust package authors to only put a package in once}
\usage{
strip_duplicate_dependencies(dependencies)
}
\description{
We don't trust package authors to only put a package in once
}

