% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependecies.R
\name{swap_packages}
\alias{swap_packages}
\title{Swap packages by comparing version information. 1ist2 has already
been ordered by dependency, so is imperative that we keep its order while
potentially swapping in the corresponding packages in list1 that have later
version requirements.}
\usage{
swap_packages(names1, names2, list1, list2)
}
\description{
Swap packages by comparing version information. 1ist2 has already
been ordered by dependency, so is imperative that we keep its order while
potentially swapping in the corresponding packages in list1 that have later
version requirements.
}

