% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dependecies.R
\name{swap_versions}
\alias{swap_versions}
\title{Swap all package information when side1 is greater than side2}
\usage{
swap_versions(names1, names2, list1, list2)
}
\description{
Swap all package information when side1 is greater than side2
}

