% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staging.R
\name{symlink_library}
\alias{symlink_library}
\title{Symlink all packages from a destination library to a target library.}
\usage{
symlink_library(src, target)
}
\arguments{
\item{src}{character. The location of the library to use for
generating symlinks}

\item{target}{character. The location of the library which will be
populated with symlinked versions of the packages in the \code{destination}
library.}
}
\description{
Symlink all packages from a destination library to a target library.
}

