% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{version_mismatch}
\alias{version_mismatch}
\title{Find packages whose version does not match the current library's version.}
\usage{
version_mismatch(package)
}
\arguments{
\item{package}{locked_package or dependency package}
}
\value{
TRUE or FALSE according as the current library's package version
  is incorrect.
}
\description{
Find packages whose version does not match the current library's version.
}

