% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependecies.R
\name{get_ordered_dependencies}
\alias{get_ordered_dependencies}
\title{Get dependencies for all elements in lock}
\usage{
get_ordered_dependencies(lock)
}
\description{
Get dependencies for all elements in lock
}

