% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dependencies.R
\name{get_ordered_dependencies}
\alias{get_ordered_dependencies}
\title{Get dependencies for all elements in lock}
\usage{
get_ordered_dependencies(lock)
}
\arguments{
\item{lock}{list.  List of locked packages}
}
\description{
Get dependencies for all elements in lock
}

