% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/library.R
\name{get_remote}
\alias{get_remote}
\title{Create remote in form devtools' remote_download  function likes.}
\usage{
get_remote(package)
}
\description{
Create remote in form devtools' remote_download  function likes.
}

