% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependecies.R
\name{get_remote_dependencies.CRAN}
\alias{get_remote_dependencies.CRAN}
\title{For packages on CRAN we will extract to a temporary directory when we
download the accurate remote DESCRIPTION file. Because these are tarballs
there is no simple way to extract only our desired file like we can with
zipfiles using the unz function.}
\usage{
\method{get_remote_dependencies}{CRAN}(package)
}
\description{
For packages on CRAN we will extract to a temporary directory when we
download the accurate remote DESCRIPTION file. Because these are tarballs
there is no simple way to extract only our desired file like we can with
zipfiles using the unz function.
}

