% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{key_generate.R}
\alias{key_generate.R}
\title{Generate a new age identity (key pair) - DEPRECATED}
\usage{
key_generate.R(keyfile = NULL)
}
\arguments{
\item{keyfile}{Character string, path where the private key will be saved.}
}
\value{
A \code{lockbox_key} object
}
\description{
Generate a new age identity (key pair) - DEPRECATED
}
\keyword{internal}
