% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependecies.R
\name{swap_versions}
\alias{swap_versions}
\title{Swap versions information when side1 is greater than side2}
\usage{
swap_versions(names1, names2, version1, version2)
}
\description{
Swap versions information when side1 is greater than side2
}

